third_party/blink/web_tests/external/wpt/webcodecs/video-decoder-no-size-in-configure.https.any.worker_vp9_p0-expected.txt

This is a testharness.js-based test.
[FAIL] Test configure() without setting width and height
  assert_equals: decoded frame width expected 320 but got 384
Harness: the test ran to completion.
